tests/models: write the model fixtures declaratively, configure them by loading
Every node was wrapped in a build_<node>(config) factory, so nothing in the
fixture read as a model — the reason was that one definition has to be
instantiated at two scales: real scale for the schedule tests and a small scale
the end-to-end test can run in seconds.

Loading gets both without a factory. A node's source now lives under model/ as
an @module class at file scope over a free 'config' name; tests/models/loader.py
executes such a source with a namespace pre-bound in its globals and returns the
loaded module, never entering sys.modules. Two loads of one file therefore give
two independent trees, verified behaviourally: attention comes out (4096, 512) at
real scale and (256, 256) at tiny in the same process, and binding weights into
one layer leaves the others unbound.

Nesting is injection rather than import: the decoder layer receives its
attention / moe modules, and the root receives the layer tuple. That also means
the root's own source is declarative — the width-varying part is the tuple the
loader passes in, not anything the class body has to compute. Per-layer
freshness stays a correctness requirement, since renamed() is a shallow copy and
a shared child would let one layer's load() clobber another's; each layer index
now loads its whole subtree fresh.

moe.py's module-level __getattr__ lazy-export hack goes away with this. It
existed because eager module-level globals shadowed a factory's own sibling
lookup, so a later build at another scale kept resolving to the stale export;
per-load module dicts remove that path entirely.

The real-scale consumers are untouched: test_moe.py and tests/schedule/* have a
zero-line diff and still import the same names.

772 passed, all four hooks clean.
