+++ tests/models/deepseek_v4_flash/model/__init__.py
@@ -1,0 +1,4 @@
+"""Declarative DeepSeek-V4-Flash model sources: ``@module`` classes at file
+scope over a free ``config`` name (and, for cross-file nesting, other
+injected names). Not importable on their own -- load with
+``tests.models.loader.load_model``."""
